fastmcp-server/pyproject.toml
[project]
name = "fastmcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.2",
    "langchain-google-genai>=2.1.10",
    "langchain-postgres>=0.0.15",
    "pgvector>=0.3.6",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
]
